+++ src/configure.ts
@@ -4,22 +4,32 @@ import { exec } from 'child_process';
 
 import { CACHE_FOLDER } from './constants';
 import { createCacheFolderIfNeeded } from './utils';
+import { debug } from './preview';
 
 interface JestPreviewConfigOptions {
-  externalCss: string[];
+  externalCss?: string[];
+  autoPreview?: boolean;
   publicFolder?: string;
 }
 
 export async function jestPreviewConfigure(
-  options: JestPreviewConfigOptions = { externalCss: [] },
+  {
+    externalCss = [],
+    autoPreview = false,
+    publicFolder,
+  }: JestPreviewConfigOptions = { externalCss: [], autoPreview: false },
 ) {
+  if (autoPreview) {
+    autoRunPreview();
+  }
+
   if (!fs.existsSync(CACHE_FOLDER)) {
     fs.mkdirSync(CACHE_FOLDER, {
       recursive: true,
     });
   }
 
-  options.externalCss?.forEach((cssFile) => {
+  externalCss?.forEach((cssFile) => {
     // Avoid name collision
     // Example: src/common/styles.css => cache-src___common___styles.css
     const delimiter = '___';
@@ -65,15 +75,60 @@ export async function jestPreviewConfigure(
     // }
   });
 
-  if (options.publicFolder) {
+  if (publicFolder) {
     createCacheFolderIfNeeded();
     fs.writeFileSync(
       path.join(CACHE_FOLDER, 'cache-public.config'),
-      options.publicFolder,
+      publicFolder,
       {
         encoding: 'utf-8',
         flag: 'w',
       },
     );
   }
+}
+
+// Omit only, skip, todo, concurrent, each. Couldn't use Omit. Just redeclare for simplicity
+type RawIt = (...args: Parameters<jest.It>) => ReturnType<jest.It>;
+
+function patchJestFunction(it: RawIt) {
+  const originalIt = it;
+  const itWithPreview: RawIt = (name, callback, timeout) => {
+    let callbackWithPreview: jest.ProvidesCallback | undefined;
+    if (!callback) {
+      callbackWithPreview = undefined;
+    } else {
+      callbackWithPreview = async function (
+        ...args: Parameters<jest.ProvidesCallback>
+      ) {
+        try {
+          // @ts-expect-error Just forward the args
+          return await callback(...args);
+        } catch (error) {
+          debug();
+          throw error;
+        }
+      };
+    }
+    return originalIt(name, callbackWithPreview, timeout);
+  };
+  return itWithPreview;
+}
+
+function autoRunPreview() {
+  const originalIt = it;
+  let itWithPreview = patchJestFunction(it) as jest.It;
+  itWithPreview.each = originalIt.each;
+  itWithPreview.only = patchJestFunction(originalIt.only) as jest.It;
+  itWithPreview.skip = originalIt.skip;
+  itWithPreview.todo = originalIt.todo;
+  itWithPreview.concurrent = patchJestFunction(
+    originalIt.concurrent,
+  ) as jest.It;
+
+  // Overwrite global it/ test
+  // Is there any use cases that `it` and `test` is undefined?
+  it = itWithPreview;
+  test = itWithPreview;
+  fit = itWithPreview.only;
 }