+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# 0.2.0
+
+## Features
+
+- Auto preview UI whenever a test fails, you don't have to call `preview.debug()` manually.
+
 # 0.1.7
 
 ## Fixes